# specs/api/api.yml
openapi: 3.0.0
info:
  title: GrowthLog API
  version: 1.0.0
  description: |
    ## Response Envelope 規則

    | 情境 | HTTP Status | Body 格式 |
    |------|------------|-----------|
    | 成功讀取 | 200 | 直接回傳資料物件，無 success 包裝 |
    | 成功寫入 | 200 | `{"success": true}` |
    | 寵物操作成功 | 200 | `{"success": true, "pet": {...}}` |
    | 任何錯誤 | 對應 4xx | `{"success": false, "error": {"violation_type": "<CODE>"}}` |

    ## violation_type → HTTP Status 對應表

    | violation_type | HTTP Status | 語意 |
    |----------------|-------------|------|
    | `UNAUTHORIZED` | 401 | 未帶 token 或 token 無效 |
    | `FORBIDDEN` | 403 | 有效 token 但無權操作他人資源 |
    | `NOT_FOUND` | 404 | 資源不存在 |
    | `MISSING_REQUIRED_FIELD` | 400 | 必填欄位缺失 |
    | `INVALID_VALUE` | 400 | 欄位值不合法 |
    | `INVALID_STATUS_TRANSITION` | 422 | 狀態機轉換不合法 |
    | `ALREADY_EXISTS` | 422 | 業務規則衝突（如寵物已存在）|

    ## 認證

    受保護端點需在 Header 帶入：Authorization: Bearer <JWT>。
    JWT 由 POST /api/auth/google 取得，以 HS256 + JWT_SECRET 環境變數簽發。
    前端將 token 存入 auth-token cookie，每次請求自動帶入。

    ## 路徑前綴

    後端實際路徑加 /api 前綴（如 /api/goals）。
    前端透過 Next.js proxy 將 /api/v1/* rewrite 至後端 /api/*。
    本文件路徑省略 /api 前綴，以描述 API 的語意介面。
servers:
  - url: http://localhost:8000/api
    description: 本地開發後端
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      summary: 查看個人資料
      responses:
        '200':
          description: 回傳個人資料
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      operationId: updateMyProfile
      summary: 更新個人資料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────
  # Learning Goals
  # ─────────────────────────────────────────
...
  /goals/{goalId}/records:
    get:
      operationId: listRecords
      summary: 查看學習紀錄
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 回傳學習紀錄列表（依建立時間遞減排序）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordListResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createRecord
      summary: 新增學習紀錄
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data: # ASM: 圖片上傳支援，與 erm.dbml image_path 對應
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /goals/{goalId}/records/{recordId}:
    patch:
      operationId: updateRecord
      summary: 編輯學習紀錄
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      operationId: deleteRecord
      summary: 刪除學習紀錄
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
        - name: recordId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'

  # ─────────────────────────────────────────
  # Mood
  # ─────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    SuccessResponse:
      description: 操作成功（HTTP 200）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessResult'
    ErrorResponse:
      description: 操作失敗，回傳對應 4xx HTTP status + success=false body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      description: |
        所有錯誤回傳格式（含 UNAUTHORIZED、FORBIDDEN、NOT_FOUND、INVALID_VALUE 等），
        回傳對應 4xx HTTP status code，body 固定為此結構。
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [violation_type]
          properties:
            violation_type:
              type: string
              description: >
                錯誤代碼：UNAUTHORIZED | NOT_FOUND | INVALID_VALUE |
                PET_ALREADY_EXISTS | MISSING_REQUIRED_FIELD
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]

    # ── Auth ────────────────────────────────
...
    UserProfile:
      type: object
      required: [user_id, display_name]
      properties:
        user_id:
          type: integer
        display_name:
          type: string
        avatar_url:
          type: string
    UpdateUserProfileRequest:
      type: object
      properties:
        display_name:
          type: string
        avatar_url:
          type: string

    # ── Goals ───────────────────────────────
...
    LearningRecord:
      type: object
      required: [record_id, title, entry_date]
      properties:
        record_id:
          type: integer
        title:
          type: string
        content:
          type: string
        image_url:
          type: string
          format: uri
        entry_date:
          type: string
          format: date
    RecordListResponse:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/LearningRecord'
    CreateRecordRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string
        image:
          type: string
          format: binary
    UpdateRecordRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        content:
          type: string

    # ── Mood ────────────────────────────────
...